[FAB-9497] Use sdk to generate new_channel transaction
Automated the usage of configtxgen.

I have added a setup script for target venv of Makefile. This file will ensure
configtxgen is in path.

Now the client has to give the path of the
configtx file and name of channel profile. Modified the corresponding
test cases and also introduced appropriate parameters in E2E_CONFIG
dictionary.

Change-Id: Ib9293c5d77328b81a3f7f471be6bba56ad9c25f0
Signed-off-by: Sushil Paneru <sushil.paneru1@gmail.com>